requirements.txt
streamlit
markdown2
langchain
langchain-community
langchain-core
langchain-openai
langchain-groq
faiss-cpu
openai
sqlalchemy
python-dotenv
pinecone-client
PyMuPDF
speechrecognition
pyttsx3
docx2txt
langchain_pinecone
langchain_huggingface
pyperclip
streamlit_javascript